+++ internal/cli/atlas/atlas.go
@@ -19,9 +19,9 @@ import (
 	"github.com/mongodb/mongocli/internal/cli/atlas/accesslists"
 	"github.com/mongodb/mongocli/internal/cli/atlas/backup"
 	"github.com/mongodb/mongocli/internal/cli/atlas/clusters"
+	"github.com/mongodb/mongocli/internal/cli/atlas/customdbroles"
 	"github.com/mongodb/mongocli/internal/cli/atlas/customdns"
 	"github.com/mongodb/mongocli/internal/cli/atlas/datalake"
-	"github.com/mongodb/mongocli/internal/cli/atlas/dbroles"
 	"github.com/mongodb/mongocli/internal/cli/atlas/dbusers"
 	"github.com/mongodb/mongocli/internal/cli/atlas/integrations"
 	"github.com/mongodb/mongocli/internal/cli/atlas/logs"
@@ -67,7 +67,7 @@ func Builder() *cobra.Command {
 		maintenance.Builder(),
 		performanceadvisor.Builder(),
 		customdns.Builder(),
-		dbroles.Builder(),
+		customdbroles.Builder(),
 	)
 
 	return cmd
