+++ internal/cli/atlas/customdbroles/custom_db_roles_test.go
@@ -14,7 +14,7 @@
 
 // +build unit
 
-package dbroles
+package customdbroles
 
 import (
 	"testing"
