+++ internal/cli/atlas/customdbroles/describe_test.go
@@ -14,7 +14,7 @@
 
 // +build unit
 
-package dbroles
+package customdbroles
 
 import (
 	"testing"
