+++ internal/cli/atlas/customdbroles/update.go
@@ -12,14 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dbroles
+package customdbroles
 
 import (
 	"errors"
 
+	"github.com/mongodb/mongocli/internal/convert"
+
 	"github.com/mongodb/mongocli/internal/cli"
 	"github.com/mongodb/mongocli/internal/config"
-	"github.com/mongodb/mongocli/internal/convert"
 	"github.com/mongodb/mongocli/internal/flag"
 	"github.com/mongodb/mongocli/internal/store"
 	"github.com/mongodb/mongocli/internal/usage"
@@ -32,12 +33,9 @@ const updateTemplate = "Custom Database Role successfully updated.\n"
 type UpdateOpts struct {
 	cli.GlobalOpts
 	cli.OutputOpts
-	action         string
-	collection     string
-	db             string
 	roleName       string
+	action         []string
 	inheritedRoles []string
-	cluster        bool
 	append         bool
 	store          store.DatabaseRoleUpdater
 }
@@ -66,25 +64,16 @@ func (opts *UpdateOpts) Run() error {
 }
 
 func (opts *UpdateOpts) newCustomDBRole(role *atlas.CustomDBRole) *atlas.CustomDBRole {
-	out := &atlas.CustomDBRole{}
-	resource := atlas.Resource{}
-	if opts.action != "" {
-		if opts.cluster {
-			resource.Cluster = &opts.cluster
-		} else {
-			resource.Collection = opts.collection
-			resource.Db = opts.db
-		}
-
-		action := atlas.Action{
-			Action:    opts.action,
-			Resources: []atlas.Resource{resource},
-		}
-
-		out.Actions = []atlas.Action{action}
+	out := &atlas.CustomDBRole{
+		Actions:        []atlas.Action{},
+		InheritedRoles: []atlas.InheritedRole{},
 	}
 
-	if opts.inheritedRoles != nil {
+	if len(opts.action) > 0 {
+		out.Actions = convert.BuildAtlasActions(opts.action)
+	}
+
+	if len(opts.inheritedRoles) > 0 {
 		out.InheritedRoles = convert.BuildAtlasInheritedRoles(opts.inheritedRoles)
 	}
 
@@ -97,22 +86,13 @@ func (opts *UpdateOpts) newCustomDBRole(role *atlas.CustomDBRole) *atlas.CustomD
 }
 
 func (opts *UpdateOpts) validate() error {
-	if opts.cluster && (opts.collection != "" || opts.db != "") {
-		return errors.New("you can't use --cluster with --db and --collection ")
-	}
-
-	if opts.action == "" && opts.inheritedRoles == nil {
+	if len(opts.action) == 0 && opts.inheritedRoles == nil {
 		return errors.New("you must provide either actions or inherited roles")
 	}
-
-	if opts.action != "" && opts.db == "" && !opts.cluster {
-		return errors.New("you must provide --db with --action")
-	}
-
 	return nil
 }
 
-// mongocli atlas dbrole(s) update roleName --action actionName --db db --collection collection --inheritedRole role@db --append
+// mongocli atlas dbrole(s) update create <roleName> --privilege action[@dbName.collection] --inheritedRole role@db --append
 func UpdateBuilder() *cobra.Command {
 	opts := &UpdateOpts{}
 	cmd := &cobra.Command{
@@ -134,10 +114,7 @@ func UpdateBuilder() *cobra.Command {
 	}
 
 	cmd.Flags().StringSliceVar(&opts.inheritedRoles, flag.InheritedRole, []string{}, usage.InheritedRoles)
-	cmd.Flags().StringVar(&opts.action, flag.Action, "", usage.Action)
-	cmd.Flags().StringVar(&opts.db, flag.Database, "", usage.DatabaseCustomRole)
-	cmd.Flags().StringVar(&opts.collection, flag.Collection, "", usage.Collection)
-	cmd.Flags().BoolVar(&opts.cluster, flag.Cluster, false, usage.CustomDBRoleCluster)
+	cmd.Flags().StringSliceVar(&opts.action, flag.Privilege, []string{}, usage.PrivilegeAction)
 	cmd.Flags().BoolVar(&opts.append, flag.Append, false, usage.Append)
 
 	cmd.Flags().StringVar(&opts.ProjectID, flag.ProjectID, "", usage.ProjectID)
