+++ internal/convert/database_user_test.go
@@ -113,48 +113,3 @@ func TestBuildOMRoles(t *testing.T) {
 		}
 	})
 }
-
-func TestBuildAtlasInheritedRoles(t *testing.T) {
-	t.Run("No database defaults to admin", func(t *testing.T) {
-		r := BuildAtlasInheritedRoles([]string{"admin"})
-		expected := []mongodbatlas.InheritedRole{
-			{
-				Role: "admin",
-				Db:   "admin",
-			},
-		}
-		if err := deep.Equal(r, expected); err != nil {
-			t.Error(err)
-		}
-	})
-
-	t.Run("should split by @", func(t *testing.T) {
-		r := BuildAtlasInheritedRoles([]string{"admin@test"})
-		expected := []mongodbatlas.InheritedRole{
-			{
-				Role: "admin",
-				Db:   "test",
-			},
-		}
-		if err := deep.Equal(r, expected); err != nil {
-			t.Error(err)
-		}
-	})
-
-	t.Run("all", func(t *testing.T) {
-		r := BuildAtlasInheritedRoles([]string{"admin@test", "something"})
-		expected := []mongodbatlas.InheritedRole{
-			{
-				Role: "admin",
-				Db:   "test",
-			},
-			{
-				Role: "something",
-				Db:   "admin",
-			},
-		}
-		if err := deep.Equal(r, expected); err != nil {
-			t.Error(err)
-		}
-	})
-}
