+++ internal/flag/flags.go
@@ -156,13 +156,11 @@ const (
 	Granularity                     = "granularity"                     // Granularity flag
 	Key                             = "key"                             // Key flag
 	Collection                      = "collection"                      // Collection flag
-	Cluster                         = "cluster"                         // Cluster flag
 	Append                          = "append"                          // Append flag
-	Action                          = "action"                          // Action flag
+	Privilege                       = "privilege"                       // Privilege flag
 	InheritedRole                   = "inheritedRole"                   // InheritedRole flag
 	CollectionName                  = "collectionName"                  // CollectionName flag
 	Database                        = "db"                              // Database flag
-	RoleName                        = "roleName"                        // RoleName flag
 	Unique                          = "unique"                          // Unique flag
 	RSName                          = "rsName"                          // RSName flag
 	Sparse                          = "sparse"                          // Sparse flag
