+++ internal/usage/usage.go
@@ -173,11 +173,9 @@ const (
 	DatabaseUser                    = "Username of a database user."
 	MonthsUntilExpiration           = "Number of months that the certificate is valid for."
 	Collection                      = "Collection name."
-	CustomDBRoleCluster             = "Set to true to indicate that the action is granted on the cluster resource."
 	Append                          = "The input action and inheritedRoles will be appended to the existing role."
 	DatabaseCustomRole              = "Database on which the action is granted."
-	CustomRoleName                  = "Name of the custom role."
-	Action                          = "Name of the privilege action."
+	PrivilegeAction                 = "List of actions per database and collection, if no database or collections is provided then cluster is assumed"
 	InheritedRoles                  = "Each object in the inheritedRoles array represents a key-value pair indicating the inherited role and the database on which the role is granted."
 	Analyzer                        = "Analyzer to use when creating the index"
 	SearchAnalyzer                  = "Analyzer to use when searching the index."
